Move requests test to a separate file
